(Hopefully) reduce complexity in api routes.
