+++ natlas-server/app/util.py
@@ -75,63 +75,3 @@ def generate_hex_16():
 
 def generate_hex_32():
 	return "%x" % random.randrange(16**32)
-
-
-def parse_ssl_data(sslcert):
-	altnames = []
-	# Parse out Subject Alternative Names because libnmap doesn't give them to us
-	for line in sslcert['output'].split('\n'):
-		if line.startswith("Subject Alternative Name:"):
-			for item in line.split(' ')[3:]:
-				altname = item.strip(',').split('DNS:')
-				if len(altname) > 1: # Prevent an indexerror in edge case
-					altnames.append(altname[1])
-
-	elements = sslcert['elements']
-	subject = elements.get('subject')
-	issuer = elements.get('issuer')
-	pubkey = elements.get('pubkey')
-	sig_alg = elements.get('sig_algo')
-	validity = elements.get('validity')
-	md5 = elements.get('md5')
-	sha1 = elements.get('sha1')
-	pem = elements.get('pem')
-
-	result = {}
-	if subject:
-		subDict = {}
-		if subject.get('commonName'):
-			subDict['commonName'] = subject.get('commonName')
-		if altnames:
-			subDict['altNames'] = altnames
-		if subDict:
-			result['subject'] = subDict
-
-	if issuer:
-		result['issuer'] = issuer
-
-	if pubkey:
-		pubkeyDict = {}
-		if pubkey.get('type'):
-			pubkeyDict['type'] = pubkey.get('type')
-		if pubkey.get('bits'):
-			pubkeyDict['bits'] = int(pubkey.get('bits'))
-		if pubkeyDict:
-			result['pubkey'] = pubkeyDict
-
-	if sig_alg:
-		result['sig_alg'] = sig_alg
-
-	if validity:
-		if validity.get('notAfter'):
-			result['notAfter'] = validity.get('notAfter')
-		if validity.get('notBefore'):
-			result['notBefore'] = validity.get('notBefore')
-	if md5:
-		result['md5'] = md5
-	if sha1:
-		result['sha1'] = sha1
-	if pem:
-		result['pem'] = pem
-
-	return result
